Don't depend on undefiend cursor behaviour
read_exact doesn't guarantee any behaviour when not able to fill the
buffer. In recent stdlib it seems like Cursor consumes the first byte,
so we must not depend on that.
The previous implementation in rust-packet now causes panics for odd
packets.

Also add a preliminary udp odd packet size test case to at least detect
the panic this triggers.